backend/requirements.txt
fastapi
uvicorn
openai
boto3
python-dotenv
requests
streamlit
python-multipart
pydantic
